misc/knowledge_bot/requirements.txt
gradio
beautifulsoup4
requests
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
langchain-core
chromadb
sentence-transformers
openai
python-dotenv
tiktoken
dashscope
edge-tts
